# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar ONGs e pets",
        "title": "OngPet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/pets": {
            "get": {
                "description": "Retorna todos os pets cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Lista todos os Pets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.PetListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria um Pet no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Cria um novo Pet",
                "parameters": [
                    {
                        "description": "Novo Pet",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreatePetInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pets/{id}": {
            "get": {
                "description": "Retorna um Pet específico pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Busca um Pet pelo ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pet"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza os dados de um Pet pelo ID (não atualiza imagens)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Atualiza um Pet existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreatePetInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "pet": {
                                    "$ref": "#/definitions/models.Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove um Pet pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Remove um Pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pets/{id}/imagens": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Faz upload de até 5 imagens para o Pet pelo ID (usando Supabase)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Adiciona imagens a um Pet existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Imagens do Pet (até 5)",
                        "name": "imagens",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "pet": {
                                    "$ref": "#/definitions/models.Pet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Pet": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "especie": {
                    "type": "string"
                },
                "formulario_id": {
                    "description": "OPCIONAL",
                    "type": "string"
                },
                "idade": {
                    "type": "integer"
                },
                "imagens": {
                    "description": "ARRAY PostgreSQL (text[])",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PetImage"
                    }
                },
                "nome": {
                    "type": "string"
                },
                "ong_id": {
                    "type": "string"
                },
                "peso": {
                    "type": "number"
                },
                "porte": {
                    "type": "string"
                },
                "raca": {
                    "type": "string"
                },
                "regiao": {
                    "type": "string"
                }
            }
        },
        "models.PetImage": {
            "type": "object",
            "properties": {
                "pet_id": {
                    "type": "string"
                },
                "position": {
                    "description": "ordem no frontend",
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "v1.CreatePetInput": {
            "type": "object",
            "properties": {
                "descricao": {
                    "type": "string"
                },
                "especie": {
                    "type": "string"
                },
                "formulario_id": {
                    "type": "string"
                },
                "idade": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "ong_id": {
                    "type": "string"
                },
                "peso": {
                    "type": "number"
                },
                "porte": {
                    "type": "string"
                },
                "raca": {
                    "type": "string"
                },
                "regiao": {
                    "type": "string"
                }
            }
        },
...
        "v1.PetListResponse": {
            "type": "object",
            "properties": {
                "dados": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Pet"
                    }
                },
                "proxima_pagina": {
                    "type": "boolean"
                },
                "total_paginas": {
                    "type": "integer"
                },
                "total_registros": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}